Refactor Audio
This commit refactors a lot of audio by fixing a lot of nitpicks here and there.

+++ app/src/main/cpp/skyline/audio/resampler.h
@@ -20,6 +20,6 @@ namespace skyline::audio {
          * @param ratio The conversion ratio needed
          * @param channelCount The amount of channels the buffer contains
          */
-        std::vector<i16> ResampleBuffer(const std::vector<i16> &inputBuffer, double ratio, int channelCount);
+        std::vector<i16> ResampleBuffer(const std::vector<i16> &inputBuffer, double ratio, u8 channelCount);
     };
 }
